+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 13.0.0
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Add techdivision/import-cli-simple#216
+* Switch to latest techdivision/import-customer 13.* version as dependency
+
 # Version 12.0.0
 
 ## Bugfixes